pinocchio

Module account_info

Data structures to represent account information.

Structs

Account 🔒

Raw account data.

AccountInfo

Wrapper struct for an Account.

Ref

Reference to account data or lamports with checked borrow rules.

RefMut

Mutable reference to account data or lamports with checked borrow rules.

Enums

BorrowState

Represents masks for borrow state of an account.

Constants

DATA_MASK 🔒

Mask representing the mutable borrow flag for data.

DATA_SHIFT 🔒

Bytes to shift to get to the borrow state of data.

GET_LEN_MASK 🔒

Mask to retrieve the original data length.

LAMPORTS_MASK 🔒

Mask representing the mutable borrow flag for lamports.

LAMPORTS_SHIFT 🔒

Bytes to shift to get to the borrow state of lamports.

MAX_PERMITTED_DATA_INCREASE

Maximum number of bytes a program may add to an account during a single top-level instruction.

SET_LEN_MASK 🔒

Mask to indicate the original data length has been set.