pinocchio::account_info

Struct Account

#[repr(C)]
pub(crate) struct Account {
    pub(crate) borrow_state: u8,
    is_signer: u8,
    is_writable: u8,
    executable: u8,
    original_data_len: u32,
    key: Pubkey,
    owner: Pubkey,
    lamports: u64,
    pub(crate) data_len: u64,
}

Raw account data.

This data is wrapped in an AccountInfo struct, which provides safe access to the data.

Fields
borrow_state: u8

Borrow state of the account data.

We reuse the duplicate flag for this. We set it to 0b0000_0000.

We use the first four bits to track state of lamport borrow

We use the second four bits to track state of data borrow

4 bit state: [1 bit mutable borrow flag | u3 immmutable borrow flag] This gives us up to 7 immutable borrows. Note that does not mean 7 duplicate account infos, but rather 7 calls to borrow lamports or borrow data across all duplicate account infos.

is_signer: u8

Indicates whether the transaction was signed by this account.

is_writable: u8

Indicates whether the account is writable.

executable: u8

Indicates whether this account represents a program.

original_data_len: u32

Accountโs original data length when it was serialized for the current program invocation.

The value of this field is lazily initialized to the current data length and the SET_LEN_MASK flag on first access. When reading this field, the flag is cleared to retrieve the original data length by using the GET_LEN_MASK mask.

Currently, this value is only used for realloc to determine if the account data length has changed from the original serialized length beyond the maximum permitted data increase.

key: Pubkey

Public key of the account.

owner: Pubkey

Program that owns this account. Modifiable by programs.

lamports: u64

The lamports in the account. Modifiable by programs.

data_len: u64

Length of the data. Modifiable by programs.