pinocchio::account_info

Struct AccountInfo

#[repr(C)]
pub struct AccountInfo {
    pub(crate) raw: *mut Account,
}

Wrapper struct for an Account.

This struct provides safe access to the data in an Account. It is also used to track borrows of the account data and lamports, given that an account can be “shared” across multiple AccountInfo instances.

Fields
raw: *mut Account

Raw (pointer to) account data.

Note that this is a pointer can be shared across multiple AccountInfo.

Implementations

impl AccountInfo

pub fn key(&self) -> &Pubkey

Public key of the account.

pub unsafe fn owner(&self) -> &Pubkey

Program that owns this account.

Safety

A reference returned by this method is invalidated when Self::assign is called.

pub fn is_signer(&self) -> bool

Indicates whether the transaction was signed by this account.

pub fn is_writable(&self) -> bool

Indicates whether the account is writable.

pub fn executable(&self) -> bool

Indicates whether this account represents a program.

Program accounts are always read-only.

pub fn data_len(&self) -> usize

Returns the size of the data in the account.

pub fn lamports(&self) -> u64

Returns the lamports in the account.

pub fn data_is_empty(&self) -> bool

Indicates whether the account data is empty.

An account is considered empty if the data length is zero.

pub fn is_owned_by(&self, program: &Pubkey) -> bool

Checks if the account is owned by the given program.

pub unsafe fn assign(&self, new_owner: &Pubkey)

Changes the owner of the account.

Safety

Using this method invalidates any reference returned by Self::owner.

pub fn is_borrowed(&self, state: BorrowState) -> bool

Return true if the account borrow state is set to the given state.

This will test both data and lamports borrow state.

pub unsafe fn borrow_lamports_unchecked(&self) -> &u64

Returns a read-only reference to the lamports in the account.

Safety

This method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Useful when an instruction has verified non-duplicate accounts.

pub unsafe fn borrow_mut_lamports_unchecked(&self) -> &mut u64

Returns a mutable reference to the lamports in the account.

Safety

This method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Useful when an instruction has verified non-duplicate accounts.

pub unsafe fn borrow_data_unchecked(&self) -> &[u8]

Returns a read-only reference to the data in the account.

Safety

This method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Useful when an instruction has verified non-duplicate accounts.

pub unsafe fn borrow_mut_data_unchecked(&self) -> &mut [u8]

Returns a mutable reference to the data in the account.

Safety

This method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Useful when an instruction has verified non-duplicate accounts.

pub fn try_borrow_lamports(&self) -> Result<Ref<'_, u64>, ProgramError>

Tries to get a read-only reference to the lamport field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn try_borrow_mut_lamports(&self) -> Result<RefMut<'_, u64>, ProgramError>

Tries to get a read only reference to the lamport field, failing if the field is already borrowed in any form.

pub fn check_borrow_lamports(&self) -> Result<(), ProgramError>

👎Deprecated since 0.8.4: Use can_borrow_lamports instead

Checks if it is possible to get a read-only reference to the lamport field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn can_borrow_lamports(&self) -> Result<(), ProgramError>

Checks if it is possible to get a read-only reference to the lamport field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn check_borrow_mut_lamports(&self) -> Result<(), ProgramError>

👎Deprecated since 0.8.4: Use can_borrow_mut_lamports instead

Checks if it is possible to get a mutable reference to the lamport field, failing if the field is already borrowed in any form.

pub fn can_borrow_mut_lamports(&self) -> Result<(), ProgramError>

Checks if it is possible to get a mutable reference to the lamport field, failing if the field is already borrowed in any form.

pub fn try_borrow_data(&self) -> Result<Ref<'_, [u8]>, ProgramError>

Tries to get a read-only reference to the data field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn try_borrow_mut_data(&self) -> Result<RefMut<'_, [u8]>, ProgramError>

Tries to get a mutable reference to the data field, failing if the field is already borrowed in any form.

pub fn check_borrow_data(&self) -> Result<(), ProgramError>

👎Deprecated since 0.8.4: Use can_borrow_data instead

Checks if it is possible to get a read-only reference to the data field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn can_borrow_data(&self) -> Result<(), ProgramError>

Checks if it is possible to get a read-only reference to the data field, failing if the field is already mutable borrowed or if 7 borrows already exist.

pub fn check_borrow_mut_data(&self) -> Result<(), ProgramError>

👎Deprecated since 0.8.4: Use can_borrow_mut_data instead

Checks if it is possible to get a mutable reference to the data field, failing if the field is already borrowed in any form.

pub fn can_borrow_mut_data(&self) -> Result<(), ProgramError>

Checks if it is possible to get a mutable reference to the data field, failing if the field is already borrowed in any form.

pub fn realloc( &self, new_len: usize, zero_init: bool, ) -> Result<(), ProgramError>

Realloc the account’s data and optionally zero-initialize the new memory.

Note: Account data can be increased within a single call by up to MAX_PERMITTED_DATA_INCREASE bytes.

Note: Memory used to grow is already zero-initialized upon program entrypoint and re-zeroing it wastes compute units. If within the same call a program reallocs from larger to smaller and back to larger again the new space could contain stale data. Pass true for zero_init in this case, otherwise compute units will be wasted re-zero-initializing.

Safety

This method makes assumptions about the layout and location of memory referenced by AccountInfo fields. It should only be called for instances of AccountInfo that were created by the runtime and received in the process_instruction entrypoint of a program.

pub fn close(&self) -> ProgramResult

Zero out the the account’s data length, lamports and owner fields, effectively closing the account.

This doesn’t protect against future reinitialization of the account since the account data will need to be zeroed out as well; otherwise the lenght, lamports and owner can be set again before the data is wiped out from the ledger using the keypair of the account being closed.

Important

The lamports must be moved from the account prior to closing it to prevent an unbalanced instruction error.

pub unsafe fn close_unchecked(&self)

Zero out the the account’s data length, lamports and owner fields, effectively closing the account.

This doesn’t protect against future reinitialization of the account since the account data will need to be zeroed out as well; otherwise the lenght, lamports and owner can be set again before the data is wiped out from the ledger using the keypair of the account being closed.

Important

The lamports must be moved from the account prior to closing it to prevent an unbalanced instruction error.

Safety

This method is unsafe because it does not check if the account data is already borrowed. It should only be called when the account is not being used.

It also makes assumptions about the layout and location of memory referenced by AccountInfo fields. It should only be called for instances of AccountInfo that were created by the runtime and received in the process_instruction entrypoint of a program.

fn data_ptr(&self) -> *mut u8

Returns the memory address of the account data.