pinocchio

Module cpi

Cross-program invocation helpers.

Structs

ReturnData

Struct to hold the return data from an invoked program.

Constants

MAX_CPI_ACCOUNTS

Maximum number of accounts that can be passed to a cross-program invocation.

MAX_RETURN_DATA

Maximum size that can be set using set_return_data.

Functions

get_return_data

Get the return data from an invoked program.

invoke

Invoke a cross-program instruction.

invoke_signed

Invoke a cross-program instruction with signatures.

invoke_signed_unchecked⚠

Invoke a cross-program instruction with signatures but don’t enforce Rust’s aliasing rules.

invoke_unchecked⚠

Invoke a cross-program instruction but don’t enforce Rust’s aliasing rules.

set_return_data

Set the running program’s return data.

slice_invoke

Invoke a cross-program instruction from a slice of AccountInfos.

slice_invoke_signed

Invoke a cross-program instruction with signatures from a slice of AccountInfos.