pinocchio::cpi

Function invoke_unchecked

pub unsafe fn invoke_unchecked(
    instruction: &Instruction<'_, '_, '_, '_>,
    accounts: &[Account<'_>],
)

Invoke a cross-program instruction but don’t enforce Rust’s aliasing rules.

This function does not check that Accounts are properly borrowable. Those checks consume CPU cycles that this function avoids.

Safety

If any of the writable accounts passed to the callee contain data that is borrowed within the calling program, and that data is written to by the callee, then Rust’s aliasing rules will be violated and cause undefined behavior.