pinocchio

Module entrypoint

Macros and functions for defining the program entrypoint and setting up global handlers.

Re-exports

pub use lazy::InstructionContext;

pub use lazy::MaybeAccount;

Modules

lazy

Defines the lazy program entrypoint and the context to access the input buffer.

Structs

NoAllocator

An allocator that does not allocate memory.

Constants

HEAP_LENGTH

Length of the heap memory region used for program heap.

HEAP_START_ADDRESS

Start address of the memory region used for program heap.

SUCCESSDeprecated

Return value for a successful program execution.

Functions

deserialize⚠

Deserialize the input arguments.

Type Aliases

ProgramResultDeprecated

The result of a program execution.