pinocchio

Macro entrypoint

macro_rules! entrypoint {
    ( $process_instruction:ident ) => { ... };
    ( $process_instruction:ident, $maximum:expr ) => { ... };
}

Declare the program entrypoint and set up global handlers.

The main difference from the standard (SDK) entrypoint macro is that this macro represents an entrypoint that does not perform allocations or copies when reading the input buffer.

This macro emits the common boilerplate necessary to begin program execution, calling a provided function to process the program instruction supplied by the runtime, and reporting its result to the runtime.

It also sets up a global allocator and panic handler, using the crate::default_allocator! and crate::default_panic_handler! macros.

The first argument is the name of a function with this type signature:

fn process_instruction(
    program_id: &Pubkey,      // Public key of the account the program was loaded into
    accounts: &[AccountInfo], // All accounts required to process the instruction
    instruction_data: &[u8],  // Serialized instruction-specific data
) -> ProgramResult;

The second (optional) argument is the maximum number of accounts that the program is expecting. A program can receive more than the specified maximum, but any account exceeding the maximum will be ignored. When the maximum is not specified, the default is 64. This is currently the maximum number of accounts that a transaction may lock in a block.

Examples

Defining an entrypoint conditional on the bpf-entrypoint feature. Although the entrypoint module is written inline in this example, it is common to put it into its own file.

#[cfg(feature = "bpf-entrypoint")]
pub mod entrypoint {

    use pinocchio::{
        account_info::AccountInfo,
        entrypoint,
        msg,
        pubkey::Pubkey,
        ProgramResult
    };

    entrypoint!(process_instruction);

    pub fn process_instruction(
        program_id: &Pubkey,
        accounts: &[AccountInfo],
        instruction_data: &[u8],
    ) -> ProgramResult {
        msg!("Hello from my program!");
        Ok(())
    }

}

Important

The panic handler set up is different depending on whether the std library is available to the linker or not. The entrypoint macro will set up a default panic “hook”, that works with the #[panic_handler] set by the std. Therefore, this macro should be used when the program or any of its dependencies are dependent on the std library.

When the program and all its dependencies are no_std, it is necessary to set a #[panic_handler] to handle panics. This is done by the crate::nostd_panic_handler macro. In this case, it is not possible to use the entrypoint macro. Use the crate::program_entrypoint! macro instead and set up the allocator and panic handler manually.