pinocchio

Macro lazy_program_entrypoint

macro_rules! lazy_program_entrypoint {
    ( $process_instruction:ident ) => { ... };
}

Declare the lazy program entrypoint.

This entrypoint is defined as lazy because it does not read the accounts upfront. Instead, it provides an InstructionContext to the access input information on demand. This is useful when the program needs more control over the compute units it uses. The trade-off is that the program is responsible for managing potential duplicated accounts and set up a global allocator and panic handler.

The usual use-case for a crate::lazy_program_entrypoint! is small programs with a single instruction. For most use-cases, it is recommended to use the crate::program_entrypoint! macro instead.

This macro emits the boilerplate necessary to begin program execution, calling a provided function to process the program instruction supplied by the runtime, and reporting its result to the runtime. Note that it does not set up a global allocator nor a panic handler.

The only argument is the name of a function with this type signature:

fn process_instruction(
   mut context: InstructionContext, // wrapper around the input buffer
) -> ProgramResult;

Example

Defining an entrypoint and making it conditional on the bpf-entrypoint feature. Although the entrypoint module is written inline in this example, it is common to put it into its own file.

#[cfg(feature = "bpf-entrypoint")]
pub mod entrypoint {

    use pinocchio::{
        default_allocator,
        default_panic_handler,
        entrypoint::InstructionContext,
        lazy_program_entrypoint,
        msg,
        ProgramResult
    };

    lazy_program_entrypoint!(process_instruction);
    default_allocator!();
    default_panic_handler!();

    pub fn process_instruction(
        mut context: InstructionContext,
    ) -> ProgramResult {
        msg!("Hello from my `lazy` program!");
        Ok(())
    }

}