pinocchio::sysvars

Module clock

Information about the network’s clock, ticks, slots, etc.

Structs

Clock

A representation of network time.

Constants

CLOCK_ID

The ID of the clock sysvar.

DEFAULT_MS_PER_SLOT

The expected duration of a slot (400 milliseconds).

DEFAULT_TICKS_PER_SECOND

The default tick rate that the cluster attempts to achieve (160 per second).

DEFAULT_TICKS_PER_SLOT

At 160 ticks/s, 64 ticks per slot implies that leader rotation and voting will happen every 400 ms. A fast voting cadence ensures faster finality and convergence

Type Aliases

Epoch

The unit of time a given leader schedule is honored.

Slot

The unit of time given to a leader for encoding a block.

UnixTimestamp

An approximate measure of real-world time.