pinocchio::sysvars

Module instructions

Structs

Instructions

IntrospectedAccountMeta

IntrospectedInstruction

Constants

INSTRUCTIONS_ID

Sysvar1nstructions1111111111111111111111111

IS_SIGNER 🔒

The bit positions for the signer flags in the AccountMeta.

IS_WRITABLE 🔒

The bit positions for the writable flags in the AccountMeta.