pinocchio::sysvars

Module rent

This account contains the current cluster rent.

This is required for the rent sysvar implementation.

Structs

Rent

Rent sysvar data

Enums

RentDue

The return value of Rent::due.

Constants

ACCOUNT_STORAGE_OVERHEAD

Account storage overhead for calculation of base rent.

DEFAULT_BURN_PERCENT

Default percentage of collected rent that is burned.

DEFAULT_EXEMPTION_THRESHOLD

Default amount of time (in years) the balance has to include rent for the account to be rent exempt.

DEFAULT_EXEMPTION_THRESHOLD_AS_U64 🔒

Default amount of time (in years) the balance has to include rent for the account to be rent exempt as a u64.

DEFAULT_LAMPORTS_PER_BYTE_YEAR

Default rental rate in lamports/byte-year.

F64_EXEMPTION_THRESHOLD_AS_U64 🔒

The u64 representation of the default exemption threshold.

RENT_ID

The ID of the rent sysvar.